fix(reasoning): suppress disabled DeepSeek scratch
