+++ vllm_mlx/routes/responses.py
@@ -94,6 +94,7 @@ from ..service.helpers import (
     _resolve_max_tokens,
     _resolve_temperature,
     _resolve_top_p,
+    _uses_deepseek_v4_reasoning,
     _validate_model_name,
     _wait_with_disconnect,
     build_extended_sampling_kwargs,
@@ -1509,7 +1510,7 @@ async def _non_stream(
         reasoning_text,
         tool_calls,
         finish_reason,
-        include_reasoning_tail=(cfg.reasoning_parser_name != "deepseek_v4"),
+        include_reasoning_tail=not _uses_deepseek_v4_reasoning(cfg),
     )
 
     openai_response = ChatCompletionResponse(
@@ -3215,7 +3216,7 @@ async def _stream_responses(
                 reasoning_text=accumulated_reasoning_text,
                 tool_calls=None,
                 finish_reason=last_finish_reason,
-                include_reasoning_tail=(cfg.reasoning_parser_name != "deepseek_v4"),
+                include_reasoning_tail=not _uses_deepseek_v4_reasoning(cfg),
             )
             if rescue_text:
                 rescue_output_index = len(completed_output)
